[Frontend] Realtime API uses Renderer
